Fix iachar() for extended ascii characters

+++ src/lfortran/semantics/ast_common_visitor.h
@@ -2324,15 +2324,15 @@ public:
     }
 
     void visit_Interface(const AST::Interface_t /*&x*/) {
-      
+
     }
-    
+
     void visit_DerivedType(const AST::DerivedType_t /*&x*/) {
-      
+
     }
 
     void visit_Enum(const AST::Enum_t /*&x*/) {
-      
+
     }
 
     ASR::ttype_t* determine_type(const Location &loc, std::string& sym,
@@ -4170,7 +4170,7 @@ public:
             std::string arg_str;
             bool is_const_value = ASRUtils::is_value_constant(arg_value, arg_str);
             if( is_const_value ) {
-                int64_t ascii_code = arg_str[0];
+                int64_t ascii_code = uint8_t(arg_str[0]);
                 iachar_value = ASRUtils::EXPR(ASR::make_IntegerConstant_t(al, x.base.base.loc,
                                 ascii_code, type));
             }
